+++ packages/terra-button/tests/wdio/button-spec.js
@@ -153,8 +153,12 @@ describe('Button', () => {
     Terra.should.beAccessible({ context: '#action' });
     Terra.should.matchScreenshot({ selector: '#action' });
 
-    Terra.should.themeEachCustomProperty('#action', {
-      '--terra-button-action-border-radius': '26px',
+    Terra.should.themeCombinationOfCustomProperties({
+      testName: 'themed',
+      properties: {
+        '--terra-button-action-border-radius': '26px',
+      },
+      selector: '#action',
     });
 
     describe('Action-Hovered', () => {
@@ -198,8 +202,12 @@ describe('Button', () => {
       Terra.should.beAccessible({ context: '#utility' });
       Terra.should.matchScreenshot({ selector: '#utility' });
 
-      Terra.should.themeEachCustomProperty('#utility', {
-        '--terra-button-utility-border-radius': '50%',
+      Terra.should.themeCombinationOfCustomProperties({
+        testName: 'themed',
+        properties: {
+          '--terra-button-utility-border-radius': '50%',
+        },
+        selector: '#utility',
       });
     });