+++ packages/terra-button/CHANGELOG.md
@@ -4,6 +4,9 @@ Changelog
 Unreleased
 ----------
 
+## Fixed
+* Use the custom `aria-label` if explicitly provided, otherwise use `text` as its fallback value for Utility button and icon-only buttons.
+
 3.11.0 - (April 16, 2019)
 ------------------
 ### Changed