+++ packages/terra-button/CHANGELOG.md
@@ -3,8 +3,11 @@ Changelog
 
 Unreleased
 ----------
+### Added
+* Added new `ghost` variant
+* Added `compact` variant docsite examples 
 
-## Fixed
+### Fixed
 * Use the custom `aria-label` if explicitly provided, otherwise use `text` as its fallback value for Utility button and icon-only buttons.
 
 3.11.0 - (April 16, 2019)